+++ src/schema.ts
@@ -281,6 +281,8 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
     "scalar": true,
     "enum": [
       "Alternative Cut",
+      "Anniversary",
+      "Assembly Cut",
       "CC",
       "collector",
       "Collector",
@@ -295,10 +297,13 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
       "Extended",
       "Fan",
       "Festival",
+      "Final Cut",
       "imax",
       "IMAX",
+      "International Cut",
       "limited",
       "Limited",
+      "Redux",
       "Remastered",
       "Restored",
       "se",
@@ -306,6 +311,7 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
       "theatrical",
       "Theatrical",
       "Ultimate",
+      "Ultimate Cut",
       "Uncensored",
       "Uncut",
       "Unrated"
@@ -416,6 +422,7 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
       "Classic",
       "Clear Art",
       "Clear Logo",
+      "Closed Caption",
       "Colorized",
       "Complet",
       "Complete",
@@ -440,6 +447,7 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
       "FHD",
       "Fix",
       "Fixed",
+      "Forced",
       "Full HD",
       "Half Over Under",
       "Half Side By Side",
@@ -509,6 +517,7 @@ export const GUESSIT_SCHEMA: Record<string, PropertySchema> = {
       "Sample",
       "Scrambled",
       "Screener",
+      "SDH",
       "SDR",
       "SECAM",
       "Side By Side",
@@ -1076,7 +1085,7 @@ export interface GuessItResult {
   crc32?: string;
   date?: string | Date;
   disc?: number | (number)[];
-  edition?: "Alternative Cut" | "CC" | "collector" | "Collector" | "Criterion" | "DC" | "ddc" | "deluxe" | "Deluxe" | "Director's Cut" | "Director's Definitive Cut" | "extended" | "Extended" | "Fan" | "Festival" | "imax" | "IMAX" | "limited" | "Limited" | "Remastered" | "Restored" | "se" | "Special" | "theatrical" | "Theatrical" | "Ultimate" | "Uncensored" | "Uncut" | "Unrated" | ("Alternative Cut" | "CC" | "collector" | "Collector" | "Criterion" | "DC" | "ddc" | "deluxe" | "Deluxe" | "Director's Cut" | "Director's Definitive Cut" | "extended" | "Extended" | "Fan" | "Festival" | "imax" | "IMAX" | "limited" | "Limited" | "Remastered" | "Restored" | "se" | "Special" | "theatrical" | "Theatrical" | "Ultimate" | "Uncensored" | "Uncut" | "Unrated")[];
+  edition?: "Alternative Cut" | "Anniversary" | "Assembly Cut" | "CC" | "collector" | "Collector" | "Criterion" | "DC" | "ddc" | "deluxe" | "Deluxe" | "Director's Cut" | "Director's Definitive Cut" | "extended" | "Extended" | "Fan" | "Festival" | "Final Cut" | "imax" | "IMAX" | "International Cut" | "limited" | "Limited" | "Redux" | "Remastered" | "Restored" | "se" | "Special" | "theatrical" | "Theatrical" | "Ultimate" | "Ultimate Cut" | "Uncensored" | "Uncut" | "Unrated" | ("Alternative Cut" | "Anniversary" | "Assembly Cut" | "CC" | "collector" | "Collector" | "Criterion" | "DC" | "ddc" | "deluxe" | "Deluxe" | "Director's Cut" | "Director's Definitive Cut" | "extended" | "Extended" | "Fan" | "Festival" | "Final Cut" | "imax" | "IMAX" | "International Cut" | "limited" | "Limited" | "Redux" | "Remastered" | "Restored" | "se" | "Special" | "theatrical" | "Theatrical" | "Ultimate" | "Ultimate Cut" | "Uncensored" | "Uncut" | "Unrated")[];
   episode?: number | (number)[];
   episode_count?: number;
   episode_details?: string | (string)[];
@@ -1087,7 +1096,7 @@ export interface GuessItResult {
   frame_rate?: string;
   language?: string | GuessitLanguage | (string | GuessitLanguage)[];
   mimetype?: "application/x-bittorrent" | "application/x-iso9660-image" | "application/x-rar-compressed" | "application/zip" | "image/gif" | "image/jpeg" | "image/png" | "text/plain" | "video/mp2t" | "video/mp4" | "video/webm" | "video/x-matroska" | "video/x-msvideo";
-  other?: "2in1" | "3D" | "asrequested" | "Audio Fixed" | "Banner" | "Bonus" | "BT.2020" | "Classic" | "Clear Art" | "Clear Logo" | "Colorized" | "Complet" | "Complete" | "CONVERT" | "Converted" | "Cover" | "Dirfix" | "Disc Art" | "DOCU" | "Documentary" | "DOKU" | "Dolby Vision" | "Dual" | "Dual Audio" | "East Coast Feed" | "Extras" | "Fan Subtitled" | "Fanart" | "Fansub" | "Fast Subtitled" | "Fastsub" | "FHD" | "Fix" | "Fixed" | "Full HD" | "Half Over Under" | "Half Side By Side" | "Hardcoded Subtitles" | "HC" | "HD" | "HDLight" | "HDR10" | "HDR10+" | "HFR" | "High Frame Rate" | "High Quality" | "High Resolution" | "HLG" | "HQ" | "HR" | "Hybrid" | "Internal" | "Landscape" | "LD" | "LDTV" | "LiNE" | "Line Audio" | "Line Dubbed" | "Logo" | "Low Definition" | "MD" | "mHD" | "Mic Dubbed" | "Micro HD" | "Mux" | "Nfofix" | "NTSC" | "Oad" | "OAR" | "Oav" | "Obfuscated" | "Ona" | "Open Matte" | "Original Animated Video" | "Original Animation DVD" | "Original Aspect Ratio" | "Original Net Animation" | "Original Video" | "OV" | "Ova" | "Over Under" | "PAL" | "postbot" | "Poster" | "Preair" | "Proof" | "Prooffix" | "Proper" | "PS Vita" | "R5" | "RC" | "Read NFO" | "Real" | "Reencoded" | "Region 5" | "Region C" | "Remux" | "Repost" | "Retail" | "Rip" | "Sample" | "Scrambled" | "Screener" | "SDR" | "SECAM" | "Side By Side" | "Standard Dynamic Range" | "Straight to Video" | "STV" | "Sync Fixed" | "Thumbnail" | "Trailer" | "UHD" | "Ultra HD" | "Upscaled" | "Variable Frame Rate" | "VFR" | "Virtual Reality" | "Vita" | "VO" | "vost" | "West Coast Feed" | "Widescreen" | "ws" | "xpost" | "XXX" | ("2in1" | "3D" | "asrequested" | "Audio Fixed" | "Banner" | "Bonus" | "BT.2020" | "Classic" | "Clear Art" | "Clear Logo" | "Colorized" | "Complet" | "Complete" | "CONVERT" | "Converted" | "Cover" | "Dirfix" | "Disc Art" | "DOCU" | "Documentary" | "DOKU" | "Dolby Vision" | "Dual" | "Dual Audio" | "East Coast Feed" | "Extras" | "Fan Subtitled" | "Fanart" | "Fansub" | "Fast Subtitled" | "Fastsub" | "FHD" | "Fix" | "Fixed" | "Full HD" | "Half Over Under" | "Half Side By Side" | "Hardcoded Subtitles" | "HC" | "HD" | "HDLight" | "HDR10" | "HDR10+" | "HFR" | "High Frame Rate" | "High Quality" | "High Resolution" | "HLG" | "HQ" | "HR" | "Hybrid" | "Internal" | "Landscape" | "LD" | "LDTV" | "LiNE" | "Line Audio" | "Line Dubbed" | "Logo" | "Low Definition" | "MD" | "mHD" | "Mic Dubbed" | "Micro HD" | "Mux" | "Nfofix" | "NTSC" | "Oad" | "OAR" | "Oav" | "Obfuscated" | "Ona" | "Open Matte" | "Original Animated Video" | "Original Animation DVD" | "Original Aspect Ratio" | "Original Net Animation" | "Original Video" | "OV" | "Ova" | "Over Under" | "PAL" | "postbot" | "Poster" | "Preair" | "Proof" | "Prooffix" | "Proper" | "PS Vita" | "R5" | "RC" | "Read NFO" | "Real" | "Reencoded" | "Region 5" | "Region C" | "Remux" | "Repost" | "Retail" | "Rip" | "Sample" | "Scrambled" | "Screener" | "SDR" | "SECAM" | "Side By Side" | "Standard Dynamic Range" | "Straight to Video" | "STV" | "Sync Fixed" | "Thumbnail" | "Trailer" | "UHD" | "Ultra HD" | "Upscaled" | "Variable Frame Rate" | "VFR" | "Virtual Reality" | "Vita" | "VO" | "vost" | "West Coast Feed" | "Widescreen" | "ws" | "xpost" | "XXX")[];
+  other?: "2in1" | "3D" | "asrequested" | "Audio Fixed" | "Banner" | "Bonus" | "BT.2020" | "Classic" | "Clear Art" | "Clear Logo" | "Closed Caption" | "Colorized" | "Complet" | "Complete" | "CONVERT" | "Converted" | "Cover" | "Dirfix" | "Disc Art" | "DOCU" | "Documentary" | "DOKU" | "Dolby Vision" | "Dual" | "Dual Audio" | "East Coast Feed" | "Extras" | "Fan Subtitled" | "Fanart" | "Fansub" | "Fast Subtitled" | "Fastsub" | "FHD" | "Fix" | "Fixed" | "Forced" | "Full HD" | "Half Over Under" | "Half Side By Side" | "Hardcoded Subtitles" | "HC" | "HD" | "HDLight" | "HDR10" | "HDR10+" | "HFR" | "High Frame Rate" | "High Quality" | "High Resolution" | "HLG" | "HQ" | "HR" | "Hybrid" | "Internal" | "Landscape" | "LD" | "LDTV" | "LiNE" | "Line Audio" | "Line Dubbed" | "Logo" | "Low Definition" | "MD" | "mHD" | "Mic Dubbed" | "Micro HD" | "Mux" | "Nfofix" | "NTSC" | "Oad" | "OAR" | "Oav" | "Obfuscated" | "Ona" | "Open Matte" | "Original Animated Video" | "Original Animation DVD" | "Original Aspect Ratio" | "Original Net Animation" | "Original Video" | "OV" | "Ova" | "Over Under" | "PAL" | "postbot" | "Poster" | "Preair" | "Proof" | "Prooffix" | "Proper" | "PS Vita" | "R5" | "RC" | "Read NFO" | "Real" | "Reencoded" | "Region 5" | "Region C" | "Remux" | "Repost" | "Retail" | "Rip" | "Sample" | "Scrambled" | "Screener" | "SDH" | "SDR" | "SECAM" | "Side By Side" | "Standard Dynamic Range" | "Straight to Video" | "STV" | "Sync Fixed" | "Thumbnail" | "Trailer" | "UHD" | "Ultra HD" | "Upscaled" | "Variable Frame Rate" | "VFR" | "Virtual Reality" | "Vita" | "VO" | "vost" | "West Coast Feed" | "Widescreen" | "ws" | "xpost" | "XXX" | ("2in1" | "3D" | "asrequested" | "Audio Fixed" | "Banner" | "Bonus" | "BT.2020" | "Classic" | "Clear Art" | "Clear Logo" | "Closed Caption" | "Colorized" | "Complet" | "Complete" | "CONVERT" | "Converted" | "Cover" | "Dirfix" | "Disc Art" | "DOCU" | "Documentary" | "DOKU" | "Dolby Vision" | "Dual" | "Dual Audio" | "East Coast Feed" | "Extras" | "Fan Subtitled" | "Fanart" | "Fansub" | "Fast Subtitled" | "Fastsub" | "FHD" | "Fix" | "Fixed" | "Forced" | "Full HD" | "Half Over Under" | "Half Side By Side" | "Hardcoded Subtitles" | "HC" | "HD" | "HDLight" | "HDR10" | "HDR10+" | "HFR" | "High Frame Rate" | "High Quality" | "High Resolution" | "HLG" | "HQ" | "HR" | "Hybrid" | "Internal" | "Landscape" | "LD" | "LDTV" | "LiNE" | "Line Audio" | "Line Dubbed" | "Logo" | "Low Definition" | "MD" | "mHD" | "Mic Dubbed" | "Micro HD" | "Mux" | "Nfofix" | "NTSC" | "Oad" | "OAR" | "Oav" | "Obfuscated" | "Ona" | "Open Matte" | "Original Animated Video" | "Original Animation DVD" | "Original Aspect Ratio" | "Original Net Animation" | "Original Video" | "OV" | "Ova" | "Over Under" | "PAL" | "postbot" | "Poster" | "Preair" | "Proof" | "Prooffix" | "Proper" | "PS Vita" | "R5" | "RC" | "Read NFO" | "Real" | "Reencoded" | "Region 5" | "Region C" | "Remux" | "Repost" | "Retail" | "Rip" | "Sample" | "Scrambled" | "Screener" | "SDH" | "SDR" | "SECAM" | "Side By Side" | "Standard Dynamic Range" | "Straight to Video" | "STV" | "Sync Fixed" | "Thumbnail" | "Trailer" | "UHD" | "Ultra HD" | "Upscaled" | "Variable Frame Rate" | "VFR" | "Virtual Reality" | "Vita" | "VO" | "vost" | "West Coast Feed" | "Widescreen" | "ws" | "xpost" | "XXX")[];
   part?: number | (number)[];
   proper_count?: number;
   release_group?: string;